pyproject.toml
[tool.poetry]
name = "text2speech"
version = "0.1.0"
description = "convert text to speech using google speech service"
authors = ["Andriy Shapochka <ashapochka@gmail.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.8"
google-cloud-texttospeech = "^2.2.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
